sources/FastAPI/requirements.txt
numpy==1.23.4
pandas==1.5.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
SQLAlchemy==1.4.39
sqlparse==0.4.3
mock==4.0.3
oauthlib==3.2.1
tortoise-orm==0.17.3
fastapi==0.87.0
uvicorn==0.20.0
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.4.0
jwt==1.3.1
scikit-learn==1.1.3
python-multipart==0.0.5
aiomysql==0.1.1
